requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
anthropic>=0.20.0
httpx>=0.27.0
python-dotenv>=1.0.0
jinja2>=3.1.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
